refactor: move main sync to internal
